+++ build.sh
@@ -97,14 +97,6 @@ build_binary () {
     echo $outfile
 }
 
-build_tools () {
-    echo "Running tool & plugin builds..."
-
-    for cmd in cat false sleep true; do
-        build_tool $cmd "tools"
-    done
-}
-
 build_tool () {
     local cmd=$1
     local subdir=$2
@@ -228,18 +220,6 @@ docker_build() {
     local build_sha=$(git rev-parse HEAD)
     local ext=$@
 
-    # build Sensu binaries for amd64 platform
-    for cmd in cat false sleep true; do
-        echo "Building tools/$cmd for linux-amd64"
-        build_tool_binary linux amd64 $cmd "tools"
-    done
-
-    # build built-in handlers for amd64 platform
-    for cmd in ${HANDLERS[@]}; do
-        echo "Building handlers/$cmd for linux-amd64"
-        build_tool_binary linux amd64 $cmd "handlers"
-    done
-
     # When publishing image, ensure that we can bundle the web UI.
     build_dashboard $ext
 
@@ -352,9 +332,6 @@ case "$cmd" in
     "build_cli")
         build_cli "${@:2}"
         ;;
-    "build_tools")
-        build_tools
-        ;;
     "dashboard")
         test_dashboard
         ;;
@@ -392,7 +369,6 @@ case "$cmd" in
         integration_test_commands
         ;;
     *)
-        build_tools
         unit_test_commands
         integration_test_commands
         build_commands